Refresh each light independently

+++ src/nutt/light.h
@@ -161,14 +161,14 @@ protected:
 
 class Light {
 public:
-	Light(size_t index, gpio_num_t switch_pin, bool switch_active_low,
+	Light(uint8_t index, gpio_num_t switch_pin, bool switch_active_low,
 		gpio_num_t relay_pin, bool relay_active_low);
 	~Light() = delete;
 
 	static constexpr const char *TAG = "nutt.Light";
 	static constexpr const size_t NUM_EP_PER_LIGHT = 5;
 
-	inline size_t index() const { return index_; }
+	inline uint8_t index() const { return index_; }
 
 	void attach(Device &device);
